remove minified js from build scripts
