--- packages/liveview/.gitignore
@@ -1,1 +1,0 @@
-/src/minified.js
